move mp.py logic into cw module to make it scriptable
